+++ sdk/monitor/azure-mgmt-monitor/azure/mgmt/monitor/v2018_09_01/models/__init__.py
@@ -13,30 +13,19 @@ from ._models_py3 import ArmRoleReceiver
 from ._models_py3 import AutomationRunbookReceiver
 from ._models_py3 import AzureAppPushReceiver
 from ._models_py3 import AzureFunctionReceiver
-from ._models_py3 import Baseline
-from ._models_py3 import BaselineMetadataValue
-from ._models_py3 import BaselineResponse
-from ._models_py3 import CalculateBaselineResponse
-from ._models_py3 import CalculateBaselineResponseStatistics
 from ._models_py3 import EmailReceiver
 from ._models_py3 import EnableRequest
 from ._models_py3 import ErrorResponse
 from ._models_py3 import ItsmReceiver
-from ._models_py3 import LocalizableString
 from ._models_py3 import LogicAppReceiver
 from ._models_py3 import Resource
 from ._models_py3 import SmsReceiver
-from ._models_py3 import TimeSeriesInformation
 from ._models_py3 import VoiceReceiver
 from ._models_py3 import WebhookReceiver
 
 
 from ._monitor_management_client_enums import (
-    ErrorType,
-    PredictionResultType,
     ReceiverStatus,
-    ResultType,
-    Sensitivity,
 )
 
 __all__ = [
@@ -47,25 +36,14 @@ __all__ = [
     'AutomationRunbookReceiver',
     'AzureAppPushReceiver',
     'AzureFunctionReceiver',
-    'Baseline',
-    'BaselineMetadataValue',
-    'BaselineResponse',
-    'CalculateBaselineResponse',
-    'CalculateBaselineResponseStatistics',
     'EmailReceiver',
     'EnableRequest',
     'ErrorResponse',
     'ItsmReceiver',
-    'LocalizableString',
     'LogicAppReceiver',
     'Resource',
     'SmsReceiver',
-    'TimeSeriesInformation',
     'VoiceReceiver',
     'WebhookReceiver',
-    'ErrorType',
-    'PredictionResultType',
     'ReceiverStatus',
-    'ResultType',
-    'Sensitivity',
 ]