+++ sdk/monitor/azure-mgmt-monitor/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Release History
 
+## 4.0.0 (2022-06-08)
+
+**Features**
+
+  - Added operation ActionGroupsOperations.begin_create_notifications_at_action_group_resource_level
+  - Added operation ActionGroupsOperations.begin_create_notifications_at_resource_group_level
+  - Added operation ActionGroupsOperations.get_test_notifications_at_action_group_resource_level
+  - Added operation ActionGroupsOperations.get_test_notifications_at_resource_group_level
+
+**Breaking changes**
+
+  - Removed operation group BaselineOperations
+  - Removed operation group MetricBaselineOperations
+
 ## 3.1.0 (2022-03-16)
 
 **Features**